Rename the `setup` subdirectory to `setup_util`.
It conflicts with `setup.py` and confuses the linters.
